feat(medication): 药品库存支持小数数量(DECIMAL)并优化家属端UI
- 药箱库存字段 total_quantity/remaining_quantity 改为 DECIMAL(12,3) 支持小数
- 新增 Flyway 迁移 V5__use_decimal_inventory_quantities.sql
- 同步更新实体、DTO、Mapper 及 Plan/MedicineBox 服务逻辑
- 家属端仪表盘/通知样式优化，新增 timeUtils 时间工具

+++ backend/src/main/java/com/example/backend/mapper/UserMedicineBoxMapper.java
@@ -7,6 +7,7 @@ import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Update;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -42,7 +43,7 @@ public interface UserMedicineBoxMapper extends BaseMapper<UserMedicineBox> {
      * @return 影响行数
      */
     @Update("UPDATE user_medicine_box SET remaining_quantity = GREATEST(0, remaining_quantity - #{amount}) WHERE id = #{boxItemId} AND remaining_quantity IS NOT NULL")
-    int deductInventory(@Param("boxItemId") Long boxItemId, @Param("amount") int amount);
+    int deductInventory(@Param("boxItemId") Long boxItemId, @Param("amount") BigDecimal amount);
 
     /**
      * 原子恢复库存（remaining_quantity + amount）
@@ -52,5 +53,5 @@ public interface UserMedicineBoxMapper extends BaseMapper<UserMedicineBox> {
      * @return 影响行数
      */
     @Update("UPDATE user_medicine_box SET remaining_quantity = remaining_quantity + #{amount} WHERE id = #{boxItemId} AND remaining_quantity IS NOT NULL")
-    int restoreInventory(@Param("boxItemId") Long boxItemId, @Param("amount") int amount);
+    int restoreInventory(@Param("boxItemId") Long boxItemId, @Param("amount") BigDecimal amount);
 }
